server/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-telegram-bot==21.3
requests==2.31.0
pydantic==2.7.0
